Relocated mission orders to MissionOrders.py
