+++ worlds/sc2/Regions.py
@@ -6,7 +6,8 @@ from .Locations import LocationData
 from .Options import get_option_value, MissionOrder, get_enabled_campaigns, campaign_depending_orders, \
     GridTwoStartPositions, static_mission_orders, dynamic_mission_orders
 from .MissionTables import MissionInfo, vanilla_mission_req_table, \
-    MissionPools, SC2Campaign, get_goal_location, SC2Mission, MissionConnection, FillMission
+    MissionPools, SC2Campaign, get_goal_location, SC2Mission, MissionConnection
+from .MissionOrders import make_gauntlet, make_blitz, make_golden_path
 from .PoolFilter import filter_missions
 from worlds.AutoWorld import World
 
@@ -393,147 +394,6 @@ def make_dynamic_mission_order(
         return make_gauntlet(num_missions)
     elif mission_order_type == MissionOrder.option_blitz:
         return make_blitz(num_missions)
-
-
-def make_golden_path(world, num_missions) -> list[FillMission]:
-    chain_name_options = ['Mar Sara', 'Char', 'Kaldir', 'Zerus', 'Skygeirr Station',
-                   'Dominion Space', 'Korhal', 'Aiur', 'Shakuras', 'Ulnar']
-    world.random.shuffle(chain_name_options)
-
-    mission_difficulties = [MissionPools.EASY, MissionPools.MEDIUM, MissionPools.HARD, MissionPools.VERY_HARD]
-    first_chain = chain_name_options.pop()
-    first_mission = FillMission(MissionPools.STARTER, [MissionConnection(-1, SC2Campaign.GLOBAL)], first_chain,
-                                     completion_critical=True)
-
-    class Campaign:
-        def __init__(self, first_mission: FillMission, missions_remaining: int):
-            self.mission_order = [first_mission]
-            self.counter = 0
-            self.last_mission_in_chain = [0]
-            self.chain_names = [first_mission.category]
-            self.missions_remaining = missions_remaining
-            self.padding = 0
-
-        def add_mission(self, chain: int, difficulty: int, required_missions: int = 0):
-            if self.missions_remaining == 0 and difficulty is not MissionPools.FINAL:
-                return
-            self.counter += 1
-            if self.mission_order[self.last_mission_in_chain[chain]].number == required_missions or required_missions <= 1:
-                required_missions = 0
-            mission_connections = [MissionConnection(self.last_mission_in_chain[chain], SC2Campaign.GLOBAL)]
-            padding = 0
-            if self.last_mission_in_chain[chain] == self.last_mission_in_chain[0]:
-                # Adding padding to the start of new chains
-                if chain != 0:
-                    padding = self.padding
-            else:
-                # Requiring main chain progress for optional chains
-                mission_connections.append(MissionConnection(self.last_mission_in_chain[0], SC2Campaign.GLOBAL))
-            self.mission_order.append(FillMission(
-                difficulty,
-                mission_connections,
-                self.chain_names[chain],
-                number=required_missions,
-                completion_critical=chain == 0,
-                ui_vertical_padding=padding
-            ))
-            self.last_mission_in_chain[chain] = self.counter
-            if chain == 0:
-                self.padding += 1
-            self.missions_remaining -= 1
-
-    campaign = Campaign(first_mission, num_missions - 2)
-    current_required_missions = 0
-    main_chain = 0
-    while campaign.missions_remaining > 0:
-        main_chain += 1
-        mission_difficulty = mission_difficulties[min(main_chain // 2, 3)]
-        if main_chain % 2 == 1:  # Adding branches
-            chains_to_make = 0 if len(campaign.chain_names) > 5 else 2 if main_chain == 1 else 1
-            for new_chain in range(chains_to_make):
-                campaign.chain_names.append(chain_name_options.pop())
-                campaign.last_mission_in_chain.append(campaign.last_mission_in_chain[0])
-        # Updating branches
-        for side_chain in range(len(campaign.chain_names) - 1, 0, -1):
-            campaign.add_mission(side_chain, mission_difficulty)
-        # Adding main path mission
-        current_required_missions = (len(campaign.mission_order) * 3) // 4
-        campaign.add_mission(0, mission_difficulty, current_required_missions)
-    campaign.add_mission(0, MissionPools.FINAL, current_required_missions)
-    return {SC2Campaign.GLOBAL: campaign.mission_order}
-
-
-def make_gauntlet(num_missions) -> list[FillMission]:
-    mission_difficulties = [MissionPools.EASY, MissionPools.MEDIUM, MissionPools.HARD, MissionPools.VERY_HARD]
-    mission_order: list[FillMission] = []
-    row_length = 7
-    rows = math.ceil(num_missions / row_length)
-    if rows == 1:
-        column_names = ["I", "II", "III", "IV", "V", "VI", "VII"][:num_missions - 1]
-        column_names.append("Final")
-    else:
-        column_names = [f'_{col + 1}' for col in range(row_length)]
-    first_mission = FillMission(MissionPools.STARTER, [MissionConnection(-1)], column_names[0],
-                                completion_critical=True)
-    mission_order.append(first_mission)
-    mission_number = 1
-    space_rows = 0
-    while mission_number < num_missions:
-        if mission_number == num_missions - 1:
-            difficulty = MissionPools.FINAL
-        else:
-            difficulty_progress = math.floor(min(mission_number * 4 / num_missions, mission_number / 3))
-            difficulty = mission_difficulties[min(difficulty_progress, 3)]
-        connection = MissionConnection(mission_number - 1)
-        mission_order.append(FillMission(
-            difficulty,
-            [connection],
-            column_names[mission_number % row_length],
-            completion_critical=True,
-            ui_vertical_padding=space_rows
-        ))
-        mission_number += 1
-        # Next row
-        if mission_number == row_length:
-            space_rows += 1
-    return {SC2Campaign.GLOBAL: mission_order}
-
-
-def make_blitz(num_missions) -> list[FillMission]:
-    mission_difficulties = [MissionPools.EASY, MissionPools.MEDIUM, MissionPools.HARD, MissionPools.VERY_HARD]
-    min_width, max_width = 2, 5
-    mission_divisor = 5
-    dynamic_width = num_missions / mission_divisor
-    width = max(min(dynamic_width, max_width), min_width)
-    middle_column = math.floor(width / 2)
-    connections = [MissionConnection(-1)]
-    mission_number = 0
-    mission_order: List[FillMission] = []
-    if num_missions % width > middle_column:
-        final_row = math.floor(num_missions / width) * width
-        final_mission_number = final_row + middle_column
-    else:
-        final_mission_number = num_missions - 1
-    while mission_number < num_missions:
-        column = mission_number % width
-        if mission_number == middle_column:
-            difficulty = MissionPools.STARTER
-        elif mission_number == final_mission_number:
-            difficulty = MissionPools.FINAL
-        else:
-            difficulty = mission_difficulties[math.floor(min(mission_number / width, 3))]
-        mission_order.append(FillMission(
-            difficulty,
-            connections,
-            f'_{column}',
-            or_requirements=True
-        ))
-        mission_number += 1
-        # Next row, requires previous row
-        if mission_number % width == 0:
-            connections = [MissionConnection(mission_number - 1 - i) for i in range(width)]
-            connections.reverse()
-    return {SC2Campaign.GLOBAL: mission_order}
 
 
 def create_structured_regions(
